refactor: make hook like plugin
