+++ package.json
@@ -50,6 +50,7 @@
     "cross-env": "7.0.3",
     "dotenv": "10.0.0",
     "fastify": "^3.24.1",
+    "fastify-plugin": "^3.0.0",
     "fastify-swagger": "^4.12.6",
     "jsonwebtoken": "^8.5.1",
     "pg": "^8.7.1",
