+++ src/app.ts
@@ -1,11 +1,11 @@
 import fastify from 'fastify';
 import pinoms from 'pino-multi-stream';
-import jwt from 'jsonwebtoken';
 import userRouter from './resources/users/user.router';
 import boardRouter from './resources/boards/board.router';
 import taskRouter from './resources/tasks/task.router';
 import loginRouter from './resources/login/login.router';
 import env from './common/config';
+import accessPlugin from './middleware/routeAccess';
 
 const streams: pinoms.Streams = [
   { stream: process.stdout }, // an "info" level destination stream
@@ -36,30 +36,30 @@ const app = fastify({
   },
 });
 
-app.addHook('preHandler', async (req, res) => {
-  const freeLinks = ['/', '/doc', '/login'];
-  const isFreeLinks = () => freeLinks.includes(req.url);
-
-  if (!isFreeLinks()) {
-    try {
-      res.log.warn({ url: req.url }, 'test');
-      const token = req.headers.authorization!.split(' ')[1];
-      res.log.warn(
-        { url: req.url },
-        `${token} ---${Object.keys(
-          req.headers
-        )}------------------------------------------------------------------------`
-      );
-      if (!token) {
-        res.status(401).send('Wrong authorization data');
-      }
-      const isVerify = jwt.verify(token, String(env.JWT_SECRET_KEY));
-      res.log.warn({ url: req.url }, `Verify: ${isVerify}`);
-    } catch (e) {
-      res.status(401).send(e);
-    }
-  }
-});
+// app.addHook('preHandler', async (req, res) => {
+//   const freeLinks = ['/', '/doc', '/login'];
+//   const isFreeLinks = () => freeLinks.includes(req.url);
+//
+//   if (!isFreeLinks()) {
+//     try {
+//       res.log.warn({ url: req.url }, 'test');
+//       const token = req.headers.authorization!.split(' ')[1];
+//       res.log.warn(
+//         { url: req.url },
+//         `${token} ---${Object.keys(
+//           req.headers
+//         )}------------------------------------------------------------------------`
+//       );
+//       if (!token) {
+//         res.status(401).send('Wrong authorization data');
+//       }
+//       const isVerify = jwt.verify(token, String(env.JWT_SECRET_KEY));
+//       res.log.warn({ url: req.url }, `Verify: ${isVerify}`);
+//     } catch (e) {
+//       res.status(401).send(e);
+//     }
+//   }
+// });
 
 // app.register(require('fastify-swagger'), {
 //   exposeRoute: true,
@@ -80,6 +80,7 @@ routes.forEach((route: any) => {
 
 app.get('/', async () => 'Service is running!');
 
+app.register(accessPlugin);
 process.on('uncaughtException', (error) => {
   logger.error(`Process on uncaughtException error = ${error.stack}`);
   // console.log(error);
